# docs/swagger.yaml
definitions:
...
  helpers.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.Organizer:
    properties:
      bank_account_name:
        type: string
      bank_account_number:
        type: string
      bank_name:
        type: string
      commission_rate_override:
        type: number
      company_name:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      legal_entity_type:
        type: string
      tax_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      verification_status:
        type: string
    type: object
  models.OrganizerVerification:
    properties:
      created_at:
        type: string
      document_type:
        type: string
      document_url:
        type: string
      id:
        type: string
      organizer_id:
        type: string
      rejection_reason:
        type: string
      reviewed_at:
        type: string
      reviewed_by:
        type: string
      status:
        type: string
    type: object
...
  organizer.CreateOrganizerRequest:
    properties:
      bank_account_name:
        type: string
      bank_account_number:
        type: string
      bank_name:
        type: string
      commission_rate_override:
        maximum: 100
        minimum: 0
        type: number
      company_name:
        maxLength: 200
        type: string
      legal_entity_type:
        type: string
      tax_id:
        type: string
      user_id:
        type: string
    required:
    - company_name
    - user_id
    type: object
  organizer.CreateOrganizerVerificationRequest:
    properties:
      document_type:
        maxLength: 50
        type: string
      document_url:
        type: string
    required:
    - document_type
    - document_url
    type: object
  organizer.UpdateOrganizerRequest:
    properties:
      bank_account_name:
        type: string
      bank_account_number:
        type: string
      bank_name:
        type: string
      commission_rate_override:
        maximum: 100
        minimum: 0
        type: number
      company_name:
        maxLength: 200
        type: string
      legal_entity_type:
        type: string
      tax_id:
        type: string
      verification_status:
        enum:
        - pending
        - in_review
        - approved
        - rejected
        type: string
    type: object
  organizer.UpdateOrganizerVerificationRequest:
    properties:
      document_type:
        maxLength: 50
        type: string
      document_url:
        type: string
      rejection_reason:
        type: string
      reviewed_at:
        type: string
      reviewed_by:
        type: string
      status:
        enum:
        - pending
        - in_review
        - approved
        - rejected
        type: string
    type: object
...
info:
  contact: {}
  description: API for the Konsera ticketing platform
  title: Konsera API
  version: "1.0"
paths:
...
  /organizers:
    get:
      description: Get all active organizers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Organizer'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: List organizers
      tags:
      - Organizers
    post:
      consumes:
      - application/json
      description: Create an organizer profile for a user
      parameters:
      - description: Organizer data
        in: body
        name: organizer
        required: true
        schema:
          $ref: '#/definitions/organizer.CreateOrganizerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Organizer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Create organizer
      tags:
      - Organizers
  /organizers/{organizer_id}:
    delete:
      parameters:
      - description: Organizer ID
        in: path
        name: organizer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete organizer
      tags:
      - Organizers
    get:
      parameters:
      - description: Organizer ID
        in: path
        name: organizer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Organizer'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get organizer
      tags:
      - Organizers
    put:
      consumes:
      - application/json
      parameters:
      - description: Organizer ID
        in: path
        name: organizer_id
        required: true
        type: string
      - description: Organizer changes
        in: body
        name: organizer
        required: true
        schema:
          $ref: '#/definitions/organizer.UpdateOrganizerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Organizer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update organizer
      tags:
      - Organizers
  /organizers/{organizer_id}/verifications:
    get:
      parameters:
      - description: Organizer ID
        in: path
        name: organizer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OrganizerVerification'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: List organizer verifications
      tags:
      - Organizer Verifications
    post:
      consumes:
      - application/json
      parameters:
      - description: Organizer ID
        in: path
        name: organizer_id
        required: true
        type: string
      - description: Verification document
        in: body
        name: verification
        required: true
        schema:
          $ref: '#/definitions/organizer.CreateOrganizerVerificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OrganizerVerification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Submit organizer verification
      tags:
      - Organizer Verifications
  /organizers/{organizer_id}/verifications/{verification_id}:
    delete:
      parameters:
      - description: Organizer ID
        in: path
        name: organizer_id
        required: true
        type: string
      - description: Verification ID
        in: path
        name: verification_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete organizer verification
      tags:
      - Organizer Verifications
    get:
      parameters:
      - description: Organizer ID
        in: path
        name: organizer_id
        required: true
        type: string
      - description: Verification ID
        in: path
        name: verification_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrganizerVerification'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get organizer verification
      tags:
      - Organizer Verifications
    put:
      consumes:
      - application/json
      parameters:
      - description: Organizer ID
        in: path
        name: organizer_id
        required: true
        type: string
      - description: Verification ID
        in: path
        name: verification_id
        required: true
        type: string
      - description: Verification changes
        in: body
        name: verification
        required: true
        schema:
          $ref: '#/definitions/organizer.UpdateOrganizerVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrganizerVerification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update organizer verification
      tags:
      - Organizer Verifications
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"